Lg/yvlasiuk

+++ karma/src/main/java/com/gliesereum/karma/controller/business/BaseBusinessController.java
@@ -1,9 +1,10 @@
 package com.gliesereum.karma.controller.business;
 
+import com.gliesereum.karma.facade.business.BusinessPermissionFacade;
 import com.gliesereum.karma.facade.client.ClientFacade;
+import com.gliesereum.karma.model.common.BusinessPermission;
 import com.gliesereum.karma.model.document.BusinessDocument;
 import com.gliesereum.karma.service.business.BaseBusinessService;
-import com.gliesereum.karma.service.business.BusinessCategoryFacade;
 import com.gliesereum.karma.service.comment.CommentService;
 import com.gliesereum.karma.service.es.BusinessEsService;
 import com.gliesereum.karma.service.media.MediaService;
@@ -46,9 +47,6 @@ public class BaseBusinessController {
     private MediaService mediaService;
 
     @Autowired
-    private BusinessCategoryFacade businessCategoryFacade;
-
-    @Autowired
     private CommentService commentService;
 
     @Autowired
@@ -56,6 +54,9 @@ public class BaseBusinessController {
 
     @Autowired
     private ClientFacade clientFacade;
+
+    @Autowired
+    private BusinessPermissionFacade businessPermissionFacade;
 
     @GetMapping
     public List<BaseBusinessDto> getAll() {
@@ -131,19 +132,19 @@ public class BaseBusinessController {
 
     @PostMapping("/media")
     public MediaDto create(@RequestBody @Valid MediaDto media) {
-        businessCategoryFacade.throwExceptionIfUserDontHavePermissionToAction(media.getObjectId());
+        businessPermissionFacade.checkPermissionByBusiness(media.getObjectId(), BusinessPermission.BUSINESS_ADMINISTRATION);
         return mediaService.create(media);
     }
 
     @PutMapping("/media")
     public MediaDto update(@RequestBody @Valid MediaDto media) {
-        businessCategoryFacade.throwExceptionIfUserDontHavePermissionToAction(media.getObjectId());
+        businessPermissionFacade.checkPermissionByBusiness(media.getObjectId(), BusinessPermission.BUSINESS_ADMINISTRATION);
         return mediaService.update(media);
     }
 
     @DeleteMapping("/{id}/media/{mediaId}")
     public MapResponse delete(@PathVariable("id") UUID businessId, @PathVariable("mediaId") UUID mediaId) {
-        businessCategoryFacade.throwExceptionIfUserDontHavePermissionToAction(businessId);
+        businessPermissionFacade.checkPermissionByBusiness(businessId, BusinessPermission.BUSINESS_ADMINISTRATION);
         mediaService.delete(mediaId, businessId);
         return new MapResponse("true");
     }
